cleanup of existing code for pull request

+++ src/main/java/org/apache/sysds/parser/BuiltinFunctionExpression.java
@@ -675,66 +675,28 @@ public class BuiltinFunctionExpression extends DataIdentifier {
 		}
 		//implement the calculation of the transformation matrix for affine transformation of images
 			case IMG_TRANSFORM_MATRIX:
-				//transformation matrix
+				//transformation matrix must be 3x3 matrix
 				Expression expressionOne_IMG = getFirstExpr();
-				//original width
+				//dimension matrix must be a 2x2 matrix
 				Expression expressionTwo_IMG = getSecondExpr();
 				checkMatrixFrameParam(expressionOne_IMG);
 				checkMatrixFrameParam(expressionTwo_IMG);
 
-                if ((expressionOne_IMG.getOutput().getDim1() != expressionOne_IMG.getOutput().getDim2()) && expressionOne_IMG.getOutput().getDim1() != 3) {
-                    raiseValidateError("The first argument to " + _opcode + " must be a square 3x3 matrix.", false, LanguageErrorCodes.INVALID_PARAMETERS);
-                } else if ((expressionTwo_IMG.getOutput().getDim1() != expressionTwo_IMG.getOutput().getDim2()) && expressionOne_IMG.getOutput().getDim1() != 2) {
-                    raiseValidateError("The second argument to " + _opcode + " must be a square 2x2 matrix.", false, LanguageErrorCodes.INVALID_PARAMETERS);
-                }
-//			else if(expressionOne.getOutput() == null || expressionOne.getOutput().getDim1() == 0 ||
-//					expressionOne.getOutput().getDim2() == 0) {
-//				raiseValidateError("The first argument to " + _opcode + " cannot be an empty matrix.", false,
-//						LanguageErrorCodes.INVALID_PARAMETERS);
-//			}
-//			else if(expressionTwo != null) {
-//				if(expressionTwo.getOutput() == null || expressionTwo.getOutput().getDim1() == 0 ||
-//						expressionTwo.getOutput().getDim2() == 0) {
-//					raiseValidateError("The second argument to " + _opcode
-//									+ " cannot be an empty matrix. Provide either only a real matrix or a filled real and imaginary one.",
-//							false, LanguageErrorCodes.INVALID_PARAMETERS);
-//				}
-//			}
-
-//			checkNumParameters(expressionTwo != null ? 2 : 1);
-//			checkMatrixParam(expressionOne);
-//			if(expressionTwo != null && expressionOne != null) {
-//				checkMatrixParam(expressionTwo);
-//				if(expressionOne.getOutput().getDim1() != expressionTwo.getOutput().getDim1() ||
-//						expressionOne.getOutput().getDim2() != expressionTwo.getOutput().getDim2())
-//					raiseValidateError("The real and imaginary part of the provided matrix are of different dimensions.",
-//							false);
-//				else if(!isPowerOfTwo(expressionTwo.getOutput().getDim2())) {
-//					raiseValidateError(
-//							"This IFFT_LINEARIZED implementation is only defined for matrices with columns that are powers of 2.",
-//							false, LanguageErrorCodes.INVALID_PARAMETERS);
-//				}
-//			}
-//			else if(expressionOne != null) {
-//				if(!isPowerOfTwo(expressionOne.getOutput().getDim2())) {
-//					raiseValidateError(
-//							"This IFFT_LINEARIZED implementation is only defined for matrices with columns that are powers of 2.",
-//							false, LanguageErrorCodes.INVALID_PARAMETERS);
-//				}
-//			}
+				if ((expressionOne_IMG.getOutput().getDim1() != expressionOne_IMG.getOutput().getDim2()) && expressionOne_IMG.getOutput().getDim1() != 3) {
+					raiseValidateError("The first argument to " + _opcode + " must be a square 3x3 matrix.", false, LanguageErrorCodes.INVALID_PARAMETERS);
+				} else if ((expressionTwo_IMG.getOutput().getDim1() != expressionTwo_IMG.getOutput().getDim2()) && expressionOne_IMG.getOutput().getDim1() != 2) {
+					raiseValidateError("The second argument to " + _opcode + " must be a square 2x2 matrix.", false, LanguageErrorCodes.INVALID_PARAMETERS);
+				}
 
 				DataIdentifier img_transfrom_matrix_Out1 = (DataIdentifier) getOutputs()[0];
 				DataIdentifier img_transfrom_matrix_Out2 = (DataIdentifier) getOutputs()[1];
 
+				//describe the matrix characteristics type and value
 				img_transfrom_matrix_Out1.setDataType(DataType.MATRIX);
 				img_transfrom_matrix_Out1.setValueType(ValueType.FP64);
-				//img_transfrom_matrix_Out1.setDimensions(getThirdExpr().getOutput()._dim1, getFourthExpr().getOutput()._dim1); //TODO: get dimensions right, extract from second input matrix
-				//out1.setBlocksize(getFirstExpr().getOutput().getBlocksize());
-
 				img_transfrom_matrix_Out2.setDataType(DataType.MATRIX);
-				img_transfrom_matrix_Out2.setValueType(ValueType.FP32);
-				//img_transfrom_matrix_Out2.setDimensions(1,1);
-
+				img_transfrom_matrix_Out2.setValueType(ValueType.FP64);
+				//the output dimensions are not known beforehand and vary based on input values
 				break;
 		case REMOVE: {
 			checkNumParameters(2);
